powerpc: Add relocation code for fixups
We need to patch an instruction that is covered by the fixup
framework. If we don't do anything about it we end up getting
our own patched instruction unpatched by nops by the fixups.

So add an export to the fixup code that allows us to tell it
that an instruction moved location in memory. This works because
we move the instruction into a different location, but still
execute it.

Signed-off-by: Alexander Graf <[hidden email]>

Source:
http://linuxppc.10917.n7.nabble.com/PATCH-0-4-powerpc-Enable-ILE-on-pSeries-without-H-MODE-SET-td78917.html