powerpc: Don't return to BE mode when we are already there
Our Little Endian kernels can now live in a world where they are
running with Big Endian interrupts enabled. That is great for kexec,
because now we don't have to switch back to Big Endian mode.

Indicate this in the code. Only try to go into Big Endian mode when
we're not already there yet.

Signed-off-by: Alexander Graf <[hidden email]>

Source:
http://linuxppc.10917.n7.nabble.com/PATCH-0-4-powerpc-Enable-ILE-on-pSeries-without-H-MODE-SET-td78917.html